+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Changed
+- Restart the broker when its SSLListener secret changes. See
+  [#317](https://github.com/astarte-platform/astarte-kubernetes-operator/issues/317).
 ### Removed
 - Remove support for Astarte <= 0.11.